+++ wemake_python_styleguide/presets/types/tree.py
@@ -84,6 +84,7 @@ PRESET: Final = (
     compares.UnaryCompareVisitor,
     compares.WrongConstantCompareVisitor,
     compares.InCompareSanityVisitor,
+    compares.WrongFloatComplexCompareVisitor,
 
     conditions.IfStatementVisitor,
     conditions.BooleanConditionVisitor,