+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Semantic versioning in our case means:
 - Forbids to use raw strings `r''` when not necessary
 - Forbids to use too complex `f`-strings
 - Forbids to use too many `raise` statements inside a single function
+- Forbids to compare with `float` and `complex` values
 
 ### Bugfixes
 
@@ -99,6 +100,7 @@ We also have this [nice migration guide](https://wemake-python-stylegui.de/en/la
 - Adds `UnreadableNameViolation` as `WPS124` because there are some
 character combination which is not easy to read
 - Adds support for `NamedExpr` with in compare type violation
+- Forbids `float` and `complex` compares
 
 ### Bugfixes